fixed issues on png generator

+++ src/data_generator/png_generator.py
@@ -39,7 +39,7 @@ class PNGGenerator(DataGenerator):
         logging.info(f"dimension of images: {dim} x {dim} x 3")
         for i in range(0, int(self.total_files_to_generate)):
             out_path_spec = self._file_list[i]
-            records = random.randint(255, size=(dim, dim, 3))
+            records = random.randint(255, size=(dim, dim, 3), dtype=np.uint8)
             img = im.fromarray(records)
             if self.my_rank == 0 and i % 100 == 0:
                 logging.info(f"Generated file {i}/{self.total_files_to_generate}")
